pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Gil <sgy@tpay.co.kr>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "2.*"
djangorestframework = "^3.12.2"
model_bakery = "^1.2.1"
ipython = "^7.19.0"
notebook = "^6.1.5"
django-extensions = "^3.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
